# Faster code execution with Numba and Numpy