--- a/requirements.txt
+++ b/requirements.txt
@@ -3,3 +3,5 @@ coverage
 nose
 pycodestyle
 twine
+numba
+numpy