@lmas 2020-07-19
Hello and thanks for your request!

Appreciate your work but unfortunately I'll have to say no, just as I've said before to others. Please see the open issue #4